automation/requirements.txt
google-generativeai>=0.8.3
python-dotenv>=1.0.1
feedparser>=6.0.11
openai>=1.40.0
psycopg[binary]>=3.2.0
